m2 compatability

+++ drlhp/custom_env.py
@@ -31,6 +31,7 @@ class CustomEnv(gym.Env):
 
     def step(self, action):
         obs, _base_reward, term, trunc, info = self.base_env.step(action)
+        obs = obs.astype(np.float32)
         reward = self.reward_predictor.get_reward(
             torch.from_numpy(obs), torch.from_numpy(action)
         )
@@ -38,11 +39,11 @@ class CustomEnv(gym.Env):
         self.observations.append(obs)
         self.actions.append(action)
         self.dones.append(term)
-        return obs, reward, term, trunc, info
+        return obs, float(reward), term, trunc, info
 
     def reset(self, seed=None, options=None):
         self.prev_obs, info = self.base_env.reset(seed=seed, options=options)
-        return self.prev_obs, info
+        return self.prev_obs.astype(np.float32), info
 
     def render(self):
         return self.base_env.render()
